Different boxes have different display drivers

+++ brightness.sh
@@ -1,14 +1,18 @@
 #!/bin/bash
 
-f="/sys/class/backlight/intel_backlight/brightness"
-#f="/sys/class/backlight/mba6x_backlight/brightness"
-cur=$(cat $f)
+for f in /sys/class/backlight/intel_backlight/brightness \
+         /sys/class/backlight/mba6x_backlight/brightness \
+         /sys/class/backlight/acpi_video0/brightness ; do
 
-if [ "$1" = "up" ]; then
-  let cur+=100
-  echo $cur > $f
-elif [ "$1" = "down" ]; then
-  let cur-=100
-  echo $cur > $f
+if [ -f ${f} ]; then
+
+  cur=$(cat $f)
+
+  if [ "$1" = "up" ]; then
+    let cur+=10
+    echo $cur > $f
+  elif [ "$1" = "down" ]; then
+    let cur-=10
+    echo $cur > $f
+  fi
 fi
-
